GG-18-12-64

+++ database/migrations/2021_12_15_195421_create_item_products_table.php
@@ -17,7 +17,7 @@ class CreateItemProductsTable extends Migration
             $table->increments('id');
             $table->integer('CID')->unsigned();
             $table->string('itemname');
-            $table->bigInteger('qty');
+            $table->double('qty',15,2);
             $table->string('image')->default('product-demo.png');
             $table->string('description')->nullable();
             $table->timestampsTz();
